Add header build dependencies
If a header changes, any .c files that depend on it should be rebuilt!
If a header changes, no .c file that does *not* depend on it should be
rebuilt. The compiler has a facility to report header dependencies, and
this commit uses that facility.

Now, after make clean; make -j; touch include/srun_stopwatch.h;, make
will not make, for instance, main.c because it does not rely on
include/srun_stopwatch.h.